requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
mysql-connector-python>=8.0.0
